+++ test/unit/test_light_player.c
@@ -1,7 +1,7 @@
 /*
  * This file is part of libskybrush.
  *
- * Copyright 2020-2025 CollMot Robotics Ltd.
+ * Copyright 2020-2026 CollMot Robotics Ltd.
  *
  * libskybrush is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
@@ -50,7 +50,7 @@ void setUp(void)
 void tearDown(void)
 {
     sb_light_player_destroy(&player);
-    sb_light_program_destroy(&program);
+    SB_DECREF_LOCAL(&program);
 }
 
 void test_get_color_at(void)